feat: add comment feature
