+++ src/components/Scene/CameraRig.tsx
