+++ src/components/Scene/SnowAccumulation.tsx
@@ -1,14 +1,14 @@
 import * as THREE from 'three';
-import snowSound from '../../public/audio/snowSound.wav';
-import { useCallback, useMemo, useRef, useState } from 'react';
+import snowSound from '/public/audio/snowSound.wav';
+import { useCallback, useMemo, useRef } from 'react';
 import { Instances } from '@react-three/drei';
 import { MathUtils } from 'three';
-import { SnowBlock } from '@/components/SnowBlock';
-import { useFrame, useThree } from '@react-three/fiber';
+import { SnowBlock } from '@/components/Scene/SnowBlock';
+import { useFrame } from '@react-three/fiber';
 
 type AccumulationProps = {
-  count?: number;
-  position: THREE.Vector3;
+	count?: number;
+	position: THREE.Vector3;
 };
 
 const SnowAccumulation = ({ count = 20, position }: AccumulationProps) => {
@@ -35,19 +35,19 @@ const SnowAccumulation = ({ count = 20, position }: AccumulationProps) => {
 		return p;
 	}, [count]);
 
-  const handlePointEnter = useCallback((e: any) => {
-    e.stopPropagation();
-    // 눈을 0.1씩 깎아 내림
-    const t = e.eventObject.position.clone();
-    e.eventObject.position.y = MathUtils.lerp(t.y, t.y - 0.2, 0.2);
-    // 눈 치우는 소리 재생
-    snowEffectSound.play();
-  }, []);
+	const handlePointEnter = useCallback((e: any) => {
+		e.stopPropagation();
+		// 눈을 0.1씩 깎아 내림
+		const t = e.eventObject.position.clone();
+		e.eventObject.position.y = MathUtils.lerp(t.y, t.y - 0.2, 0.2);
+		// 눈 치우는 소리 재생
+		snowEffectSound.play();
+	}, []);
 
 	return (
 		<Instances ref={ref} position={position} limit={count} range={count}>
 			<boxGeometry />
-			{points.map((pt, i) => ( 
+			{points.map((pt, i) => (
 				<SnowBlock
 					scale={0.3}
 					onPointerDown={(e) => handlePointEnter(e)}
