+++ src/components/Scene/SnowBlock.tsx
