+++ src/components/Scene/SnowScene.tsx
